requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
xgboost>=1.7.0
torch>=2.0.0
pennylane>=0.32.0
qiskit>=0.44.0
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0
notebook>=7.0.0
pytest>=7.4.0
imbalanced-learn>=0.11.0
